+++ client/src/components/posts/PostStep.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { deleteStep } from '../../actions/steps-tagsActions';
 import { Card, CardText, CardImg, Button } from 'reactstrap';
-import "../../postSteps.scss";
+import "./postSteps.scss";
 
 class PostStep extends React.Component {
   state = {
