+++ client/src/components/posts/post.scss
